# refactor: remove channels where unnecessary