--- a/agent/sampler.go
+++ b/agent/sampler.go
@@ -1,8 +1,6 @@
 package main
 
 import (
-	"sync"
-
 	log "github.com/cihub/seelog"
 
 	"github.com/DataDog/datadog-trace-agent/config"
@@ -13,11 +11,7 @@ import (
 
 // Sampler chooses wich spans to write to the API
 type Sampler struct {
-	in  chan model.Trace
-	out chan []model.Trace
-
 	sampledTraces []model.Trace
-	mu            sync.Mutex
 
 	// statistics
 	traceCount int
@@ -29,14 +23,12 @@ type Sampler struct {
 type SamplerEngine interface {
 	Run()
 	Stop()
-	Sample(t model.Trace) bool
+	Sample(t model.Trace, root *model.Span, env string) bool
 }
 
 // NewSampler creates a new empty sampler ready to be started
-func NewSampler(in chan model.Trace, conf *config.AgentConfig) *Sampler {
+func NewSampler(conf *config.AgentConfig) *Sampler {
 	return &Sampler{
-		in:            in,
-		out:           make(chan []model.Trace),
 		sampledTraces: []model.Trace{},
 		traceCount:    0,
 		samplerEngine: sampler.NewSampler(conf.ExtraSampleRate, conf.MaxTPS),
@@ -46,41 +38,31 @@ func NewSampler(in chan model.Trace, conf *config.AgentConfig) *Sampler {
 // Run starts sampling traces
 func (s *Sampler) Run() {
 	go s.samplerEngine.Run()
-
-	for trace := range s.in {
-		if len(trace) == 1 && trace[0].IsFlushMarker() {
-			traces := s.Flush()
-			statsd.Client.Count("trace_agent.sampler.trace.kept", int64(len(traces)), nil, 1)
-			statsd.Client.Count("trace_agent.sampler.trace.total", int64(s.traceCount), nil, 1)
-			log.Debugf("flushed %d sampled traces out of %v", len(traces), s.traceCount)
-
-			s.traceCount = 0
-			s.out <- traces
-		} else {
-			s.AddTrace(trace)
-			s.traceCount++
-		}
-	}
-
-	close(s.out)
-	s.samplerEngine.Stop()
 }
 
-// AddTrace samples a trace then keep it until the next flush
-func (s *Sampler) AddTrace(trace model.Trace) {
-	if s.samplerEngine.Sample(trace) {
-		s.mu.Lock()
-		s.sampledTraces = append(s.sampledTraces, trace)
-		s.mu.Unlock()
+// Add samples a trace then keep it until the next flush
+func (s *Sampler) Add(t processedTrace) {
+	s.traceCount++
+	if s.samplerEngine.Sample(t.Trace, t.Root, t.Env) {
+		s.sampledTraces = append(s.sampledTraces, t.Trace)
 	}
+}
+
+// Stop stops the sampler
+func (s *Sampler) Stop() {
+	s.samplerEngine.Stop()
 }
 
 // Flush returns representative spans based on GetSamples and reset its internal memory
 func (s *Sampler) Flush() []model.Trace {
-	s.mu.Lock()
-	samples := s.sampledTraces
+	traces := s.sampledTraces
 	s.sampledTraces = []model.Trace{}
-	s.mu.Unlock()
+	traceCount := s.traceCount
+	s.traceCount = 0
 
-	return samples
+	statsd.Client.Count("trace_agent.sampler.trace.kept", int64(len(traces)), nil, 1)
+	statsd.Client.Count("trace_agent.sampler.trace.total", int64(traceCount), nil, 1)
+	log.Debugf("flushed %d sampled traces out of %v", len(traces), traceCount)
+
+	return traces
 }